Step 9.16: Extract EmptyBinConfirmModal, RecurringDeleteModal, EditRecurrenceModal
